// src/app/Models/PackageModel/itinerary-model.ts
export class ItineraryModel {
    ItineraryID: number;
    PackageID: number;
    DayNumber: number;
    CityID: number;
    City: string;
    HotelID: number;
    Hotel: string;
    HotelImage: string;
    RoomType: string;
    RoomPrice: number;
    TransportationProviderName: string;
    TransportationMode: string;
    TransportationPrice: number;
    ActivityDetails: string;
    CustomizedRoomPrice: number;
    CustomizedTransportationPrice: number;
    RoomPriceID: number;
    CustomizedRoomPriceID: number;
    TransportationPriceID: number;
    CustomizedTransportationPriceID: number;
}